[S05P12D103-87] back / [API] 회원 관리 요청 : 회원가입

+++ backend/src/main/java/com/ssafy/db/entity/BaseEntity.java
@@ -17,5 +17,5 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Integer id = null;
+    Long id = null;
 }
